Only output a new-line if actually asking for input
using this mapping:
```viml
  nnoremap <silent> \h :let @h=helplink#link('','')<cr>
```
If you use this mapping this will output an empty line for
certain helptags, causing a "Press enter to continue" even if nothing
has been echo'ed. So only do this if input is actually required.
Try this mapping after `:h :s`

When input is required, it will correctly ask the user for the tag to
use.